Shuffle files logging code lives in.
